Refactor to move robot model to fshare. Small bits of cleanup too
