+++ src/control/modules/motion-control/PidMotionController.hpp
@@ -1,0 +1,134 @@
+#pragma once
+
+#include <array>
+#include <rc-fshare/pid.hpp>
+#include <rc-fshare/robot_model.hpp>
+#include "FPGA.hpp"
+
+/**
+ * Robot controller that runs a PID loop on each of the four wheels.
+ */
+class PidMotionController {
+public:
+    PidMotionController() {
+        setPidValues(3.0, 0, 0, 50, 0);
+    }
+
+    void setPidValues(float p, float i, float d, unsigned int windup,
+                      float derivAlpha) {
+        for (Pid& ctl : _controllers) {
+            ctl.kp = p;
+            ctl.ki = i;
+            ctl.kd = d;
+            ctl.setWindup(windup);
+            ctl.derivAlpha = derivAlpha;
+        }
+    }
+
+    void updatePValues(float p) {
+        for (Pid& ctl : _controllers) {
+            ctl.kp = p;
+        }
+    }
+
+    void updateIValues(float i) {
+        for (Pid& ctl : _controllers) {
+            ctl.ki = i;
+        }
+    }
+
+    void updateDValues(float d) {
+        for (Pid& ctl : _controllers) {
+            ctl.kd = d;
+        }
+    }
+
+    void setTargetVel(Eigen::Vector3f target) { _targetVel = target; }
+
+    /**
+     * Return the duty cycle values for the motors to drive at the target
+     * velocity.
+     *
+     * @param encoderDeltas Encoder deltas for the four drive motors
+     * @param dt Time in ms since the last calll to run()
+     *
+     * @return Duty cycle values for each of the 4 motors
+     */
+    std::array<int16_t, 4> run(const std::array<int16_t, 4>& encoderDeltas,
+                               float dt, Eigen::Vector4d* errors = nullptr,
+                               Eigen::Vector4d* wheelVelsOut = nullptr,
+                               Eigen::Vector4d* targetWheelVelsOut = nullptr) {
+        // convert encoder ticks to rad/s
+        Eigen::Vector4d wheelVels;
+        wheelVels << encoderDeltas[0], encoderDeltas[1], encoderDeltas[2],
+            encoderDeltas[3];
+        wheelVels *= 2.0 * M_PI / ENC_TICKS_PER_TURN / dt;
+
+        Eigen::Vector4d targetWheelVels =
+            RobotModelControl.BotToWheel * _targetVel.cast<double>();
+
+        if (targetWheelVelsOut) {
+            *targetWheelVelsOut = targetWheelVels;
+        }
+
+        Eigen::Vector4d wheelVelErr = targetWheelVels - wheelVels;
+
+        if (errors) {
+            *errors = wheelVelErr;
+        }
+
+        if (wheelVelsOut) {
+            *wheelVelsOut = wheelVels;
+        }
+
+        // Calculated by checking for slippage at max accel, and decreasing
+        // appropriately
+        // Binary search works really well in this case
+        // Caution: This is dependent on the PID values so increasing the
+        // agressiveness of that will change this
+        double max_error = 3.134765625;
+        double scale = 1;
+
+        for (int i = 0; i < 4; i++) {
+            if (abs(wheelVelErr[i]) > max_error) {
+                scale = max(scale, abs(wheelVelErr[i]) / max_error);
+            }
+        }
+
+        wheelVelErr /= scale;
+        targetWheelVels = wheelVels + wheelVelErr;
+
+        std::array<int16_t, 4> dutyCycles;
+        for (int i = 0; i < 4; i++) {
+            // float dc;
+            float dc =
+                targetWheelVels[i] * RobotModelControl.DutyCycleMultiplier +
+                copysign(4, targetWheelVels[i]);
+
+            dc += _controllers[i].run(wheelVelErr[i]);
+
+            if (std::abs(dc) > FPGA::MAX_DUTY_CYCLE) {
+                // Limit to max duty cycle
+                dc = copysign(FPGA::MAX_DUTY_CYCLE, dc);
+                // Conditional integration indicating open loop control
+                _controllers[i].set_saturated(true);
+            } else {
+                _controllers[i].set_saturated(false);
+            }
+
+            dutyCycles[i] = (int16_t)dc;
+        }
+
+        return dutyCycles;
+    }
+
+    // 2048 ticks per turn. Theres is a 3:1 gear ratio between the motor and the
+    // wheel.
+    static const uint16_t ENC_TICKS_PER_TURN = 2048 * 3;
+
+private:
+    /// controllers for each wheel
+    std::array<Pid, 4> _controllers{};
+
+    Eigen::Vector3f _targetVel{};
+};
